/rank port part 2

+++ ElevatorBot/misc/formating.py
@@ -55,7 +55,7 @@ def split_into_chucks_of_max_2000_characters(text_str: str = None, text_list: li
         return return_list
 
 
-def format_timedelta(seconds: int | datetime.timedelta) -> str:
+def format_timedelta(seconds: float | datetime.timedelta) -> str:
     """Returns a formatted message that is displayed whenever a command wants to display a duration"""
 
     if isinstance(seconds, datetime.timedelta):
